Fix passing of the input structure to QE sub-workchains
When replacing the bare QE calculations with workchains, their
input was changed s.t. the structure needed to be passed to each
part. This breaks the strained workflow, because it modifies the
structure before passing it on.
In this fix, we explicitly pass on the structure to the QE sub-
workflows.