Add 'UNK*' to ADDITIONAL_RETRIEVE_LIST for pw2wannier90
The UNK files are only produced if `write_unk` is explicitly set
to `True` in the pw2wannier90 input. As such, makes sense to then
also _retrieve_ them, because we use the retrieved folder as input
to Wannier90, not the remote one.
If no UNK files are created, this does not produce an error.